+++ src/app/components/gallery/application-action/choose-action.html
