+++ src/app/components/gallery/application-list/gallery-application-list.scss
