+++ src/app/components/gallery/gallery.module.ts
@@ -3,23 +3,25 @@ import { FileTypeAssociationService } from "@batch-flask/ui/file/file-viewer/fil
 import { commonModules } from "app/common";
 import { DataSharedModule } from "app/components/data/shared";
 import { JobActionModule } from "app/components/job/action";
-import { ChooseActionComponent } from "app/components/market/application-action";
-import { MarketComponent } from "app/components/market/home";
-import {
-    ParameterInputComponent, SubmitMarketApplicationComponent, SubmitNcjTemplateComponent,
-} from "app/components/market/submit";
 import { TaskBaseModule } from "app/components/task/base";
+import { ChooseActionComponent } from "./application-action";
 import { GalleryApplicationListComponent } from "./application-list";
 import { GalleryRoutingModule } from "./gallery-routing.module";
+import { GalleryComponent } from "./home";
 import { RecentTemplateListComponent } from "./home/recent-template-list";
 import { LocalTemplateExplorerModule } from "./local-template-explorer";
 import { NcjTemplateViewerComponent } from "./ncj-template-viewer";
+import {
+    ParameterInputComponent,
+    SubmitMarketApplicationComponent,
+    SubmitNcjTemplateComponent,
+} from "./submit";
 import { SubmitLocalTemplateComponent } from "./submit-local-template";
 import { SubmitRecentTemplateComponent } from "./submit-recent-template";
 
 const components = [
     ChooseActionComponent,
-    MarketComponent,
+    GalleryComponent,
     SubmitMarketApplicationComponent,
     SubmitRecentTemplateComponent,
     SubmitNcjTemplateComponent,
