+++ src/app/components/gallery/home/gallery.html
