+++ src/app/components/gallery/home/gallery.scss
@@ -1,6 +1,6 @@
 @import "app/styles/variables";
 
-bl-market {
+bl-gallery {
     width: 100%;
     height: $contentview-height;
     $header-height: 50px;
