+++ src/app/components/gallery/home/index.ts
@@ -1,0 +1,1 @@
+export * from "./gallery.component";
