+++ src/app/components/gallery/home/recent-template-list/index.ts
