+++ src/app/components/gallery/home/recent-template-list/recent-template-list.scss
