+++ src/app/components/gallery/local-template-explorer/local-template-explorer.html
