+++ src/app/components/gallery/ncj-template-viewer/index.ts
