+++ src/app/components/gallery/ncj-template-viewer/ncj-template-viewer.component.ts
