+++ src/app/components/gallery/submit-local-template/index.ts
