+++ src/app/components/gallery/submit-local-template/submit-local-template.html
