+++ src/app/components/gallery/submit-recent-template/index.ts
