+++ src/app/components/gallery/submit-recent-template/submit-recent-template.scss
