+++ src/app/components/gallery/submit/parameter-input/parameter-input.scss
