+++ src/app/components/gallery/submit/submit-market-application.scss
