+++ src/app/components/gallery/submit/submit-ncj-template.component.ts
