+++ src/app/components/gallery/submit/submit-ncj-template.scss
