--- src/app/components/market/home/index.ts
@@ -1,1 +1,0 @@
-export * from "./market.component";
